refactor: upload controller
